+++ Notes/ContentViewModel.swift
@@ -15,12 +15,12 @@ class ContentViewModel: ObservableObject {
     @Published var isLinkActive = false
     
     let footnoteFont = Font.system(.footnote, design: .monospaced)
-    //let presistenceController = PersistenceController()
+    
     
     func usernameEqualToInput() {
         let user = fetchUserByUsername(inputUsername: usernameInput)
         if(user != nil && user?.password == passwordInput) {
-            isLinkActive = true 
+            isLinkActive = true
             errorMessage = ""
         }
         else {
@@ -29,7 +29,7 @@ class ContentViewModel: ObservableObject {
     }
     
     func fetchUserByUsername(inputUsername: String) -> User? {
-		let user = PersistenceController.shared.fetchUsersByUsername(username: usernameInput)
+        let user = DataManager.shared.fetchUsersByUsername(username: inputUsername)
         return user
     }
