--- Notes/ViewModifier/UnderlineTextField.swift
@@ -1,18 +1,0 @@
-//
-//  UnderlineTextField.swift
-//  Notes
-//
-//  Created by Lena Vadakkel on 24.11.23.
-//
-
-import SwiftUI
-
-struct UnderlineTextField: ViewModifier {
-	func body(content: Content) -> some View {
-		content
-			.padding(.vertical, 10)
-			.overlay(Rectangle().frame(height: 2).padding(.top, 35))
-			.foregroundColor(Color("AccentColor"))
-			.padding(10)
-	}
-}
