--- Notes/en.lproj/Localizable.strings
@@ -1,13 +1,0 @@
-/* 
-  Localizable.strings
-  Notes
-
-  Created by Lena Vadakkel on 24.11.23.
-  
-*/
-
-"Sign In" = "Sign In";
-"Sign Up" = "Sign Up";
-"Username" = "Username";
-"Password" = "Password";
-"Email" = "Email";
